+++ src/impulse/cli.py
@@ -27,12 +27,21 @@ def main():
         "and display them as dashed lines."
     ),
 )
+@click.option(
+    "--depth",
+    type=click.IntRange(min=0),
+    default=1,
+    help="Depth of module hierarchy to visualize.",
+)
 @click.argument("module_name", type=str)
-def drawgraph(module_name: str, show_import_totals: bool, show_cycle_breakers: bool) -> None:
+def drawgraph(
+    module_name: str, show_import_totals: bool, show_cycle_breakers: bool, depth: int
+) -> None:
     use_cases.draw_graph(
         module_name=module_name,
         show_import_totals=show_import_totals,
         show_cycle_breakers=show_cycle_breakers,
+        depth=depth,
         sys_path=sys.path,
         current_directory=os.getcwd(),
         get_top_level_package=adapters.get_top_level_package,